+++ packages/plugin-sound-cloud/src/soundCloud.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { findDOMNode } from 'react-dom';
-import { mergeStyles } from 'wix-rich-content-common';
+import { mergeStyles, Context } from 'wix-rich-content-common';
 import SoundCloudViewer from './soundCloud-viewer';
 import styles from '../statics/styles/default-sound-cloud-styles.scss';
 import { SOUND_CLOUD_TYPE } from './types';
@@ -29,7 +29,6 @@ class SoundCloud extends Component {
       isLoaded: false,
       isPlayable,
     };
-    this.styles = mergeStyles({ styles, theme: this.props.theme });
   }
 
   setPlayer = player => {
@@ -108,8 +107,8 @@ class SoundCloud extends Component {
   };
 
   render() {
-    const { styles } = this;
-    const { className, onClick, t } = this.props;
+    this.styles = this.styles || mergeStyles({ styles, theme: this.context.theme });
+    const { className, onClick } = this.props;
     const { isPlayable } = this.state;
     const containerClassNames = classNames(styles.soundCloud_container, className || '');
     /* eslint-disable jsx-a11y/no-static-element-interactions */
@@ -120,13 +119,15 @@ class SoundCloud extends Component {
         className={containerClassNames}
         onKeyDown={e => this.onKeyDown(e, onClick)}
       >
-        {!isPlayable && this.renderOverlay(styles, t)}
+        {!isPlayable && this.renderOverlay(styles, this.context.t)}
         {this.renderPlayer()}
       </div>
     );
     /* eslint-enable jsx-a11y/no-static-element-interactions */
   }
 }
+
+SoundCloud.contextType = Context.type;
 
 SoundCloud.propTypes = {
   componentData: PropTypes.object.isRequired,
@@ -135,8 +136,6 @@ SoundCloud.propTypes = {
   blockProps: PropTypes.object.isRequired,
   onClick: PropTypes.func.isRequired,
   className: PropTypes.string.isRequired,
-  theme: PropTypes.object.isRequired,
-  t: PropTypes.func,
 };
 
 export { SoundCloud as Component, DEFAULTS };