+++ packages/plugin-video/src/video-component.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { findDOMNode } from 'react-dom';
-import { mergeStyles } from 'wix-rich-content-common';
+import { mergeStyles, Context } from 'wix-rich-content-common';
 import VideoViewer from './video-viewer';
 import styles from '../statics/styles/default-video-styles.scss';
 import { VIDEO_TYPE_LEGACY, VIDEO_TYPE } from './types';
@@ -27,7 +27,6 @@ class VideoComponent extends React.Component {
       isLoaded: false,
       isPlayable,
     };
-    this.styles = mergeStyles({ styles, theme: this.props.theme });
   }
 
   setPlayer = player => {
@@ -107,8 +106,8 @@ class VideoComponent extends React.Component {
   };
 
   render() {
-    const { styles } = this;
-    const { className, onClick, t } = this.props;
+    this.styles = this.styles || mergeStyles({ styles, theme: this.context.theme });
+    const { className, onClick } = this.props;
     const { isPlayable } = this.state;
     const containerClassNames = classNames(styles.video_container, className || '');
     /* eslint-disable jsx-a11y/no-static-element-interactions */
@@ -119,7 +118,7 @@ class VideoComponent extends React.Component {
         className={containerClassNames}
         onKeyDown={e => this.onKeyDown(e, onClick)}
       >
-        {!isPlayable && this.renderOverlay(styles, t)}
+        {!isPlayable && this.renderOverlay(styles, this.context.t)}
         {this.renderPlayer()}
       </div>
     );
@@ -135,8 +134,8 @@ VideoComponent.propTypes = {
   blockProps: PropTypes.object.isRequired,
   onClick: PropTypes.func.isRequired,
   className: PropTypes.string.isRequired,
-  theme: PropTypes.object.isRequired,
-  t: PropTypes.func,
 };
+
+VideoComponent.contextType = Context.type;
 
 export { VideoComponent as Component, DEFAULTS };